+++ Changes.md
@@ -29,6 +29,7 @@ Improvements
 - PathListingWidget : Improved formatting of Box and Matrix values.
 - LocaliseAttributes : Added support for localising global attributes, controlled by the new `includeGlobalAttributes` plug.
 - AttributeTweaks : Added support for localising global attributes when `localise` is enabled.
+- AttributeQuery : Added support for querying global attributes when `inherit` is enabled.
 
 Fixes
 -----
@@ -93,6 +94,7 @@ Breaking Changes
 - ContextAlgo : Removed deprecated API. Use ScriptNodeAlgo instead, which has been available from Gaffer 1.4.13.0 onwards.
 - ScriptNodeAlgo : Reimplemented using Metadata rather than Context variables for storage. Use the ScriptNodeAlgo API instead of attempting direct access to `ui:*` context variables.
 - AttributeTweaks : Tweaks with `localise` enabled and a mode of `CreateIfMissing` will now not create an attribute if it is missing from the scene hierarchy, but exists in the globals.
+- AttributeQuery : Queries with `inherit` enabled will now return a result when querying an attribute that does not exist in the scene hierarchy, but does exist in the globals.
 
 Build
 -----